api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
pydantic==2.7.1
sqlalchemy==2.0.30
cryptography==42.0.7
pyjwt==2.8.0
bcrypt==4.1.3
redis==5.0.4
celery==5.4.0
celery-redbeat==2.2.0
selenium==4.21.0
numpy==1.26.4
pillow==10.3.0